riscv: Add inline IRQ process functions
Remove functions from RISC-V chips.
